Consolidated tests into one file per code review.
