+++ src/applications/vaos/tests/new-appointment/components/ConfirmationPage/index.unit.spec.jsx
@@ -28,12 +28,26 @@ describe('VAOS <ConfirmationPage>', () => {
     const flowType = FLOW_TYPES.DIRECT;
     const data = {
       typeOfCareId: '323',
+      phoneNumber: '1234567890',
+      email: 'joeblow@gmail.com',
+      reasonForAppointment: 'routine-follow-up',
+      reasonAdditionalInfo: 'Additional info',
       vaFacility: '983',
+    };
+    const facilityDetails = {
+      id: 'var983',
+      name: 'Cheyenne VA Medical Center',
+      address: {
+        postalCode: '82001-5356',
+        city: 'Cheyenne',
+        state: 'WY',
+        line: ['2360 East Pershing Boulevard'],
+      },
     };
 
     const screen = renderWithStoreAndRouter(
       <ConfirmationPage
-        fetchFacilityDetails={fetchFacilityDetails}
+        facilityDetails={facilityDetails}
         flowType={flowType}
         data={data}
         slot={{ start, end }}
@@ -45,20 +59,53 @@ describe('VAOS <ConfirmationPage>', () => {
       },
     );
 
-    screen.getByText(/Your appointment has been scheduled./i);
-    screen.getByText(moment(start).format('MMMM D, YYYY [at] h:mm a CT'));
+    expect(screen.getByText(/Your appointment has been scheduled./i)).to.be.ok;
+    expect(
+      screen.getByText(moment(start).format('MMMM D, YYYY [at] h:mm a CT')),
+    ).to.be.ok;
+
+    expect(screen.getByText(/Cheyenne VA Medical Center/i)).to.be.ok;
+    expect(screen.getByText(/2360 East Pershing Boulevard/i)).to.be.ok;
+    expect(screen.getByText(/Cheyenne, WY 82001-5356/i)).to.be.ok;
+    expect(screen.getByText(/Follow-up\/Routine/i)).to.be.ok;
+    expect(screen.getByText(/Additional info/i)).to.be.ok;
+
+    expect(
+      screen.getByRole('link', {
+        name: moment(start).format(
+          '[Add] MMMM D, YYYY [appointment to your calendar]',
+        ),
+      }),
+    ).to.be.ok;
   });
 
-  it('should render appointment request view', async () => {
+  it('should render VA appointment request view', async () => {
     const flowType = FLOW_TYPES.REQUEST;
     const data = {
       typeOfCareId: '323',
+      bestTimeToCall: {
+        morning: true,
+        afternoon: true,
+      },
+      calendarData: {
+        selectedDates: [{ date: '2019-12-20', optionTime: 'AM' }],
+      },
       vaFacility: '983',
+    };
+    const facilityDetails = {
+      name: 'CHYSHR-Sidney VA Clinic',
+      id: 'var983',
+      address: {
+        postalCode: '82001-5356',
+        city: 'Cheyenne',
+        state: 'WY',
+        line: ['2360 East Pershing Boulevard'],
+      },
     };
 
     const screen = renderWithStoreAndRouter(
       <ConfirmationPage
-        fetchFacilityDetails={fetchFacilityDetails}
+        facilityDetails={facilityDetails}
         flowType={flowType}
         data={data}
       />,
@@ -68,7 +115,247 @@ describe('VAOS <ConfirmationPage>', () => {
       },
     );
 
-    screen.getByText(/Your appointment request has been submitted./i);
+    expect(screen.getByText(/Your appointment request has been submitted./i)).to
+      .be.ok;
+    expect(screen.getByText(/VA appointment/i)).to.be.ok;
+    expect(screen.getByText(/Primary care appointment/i)).to.be.ok;
+    expect(screen.getByText(/Pending/i)).to.be.ok;
+    expect(screen.getByText(/CHYSHR-Sidney VA Clinic/i)).to.be.ok;
+    expect(screen.getByText(/Cheyenne, WY/i)).to.be.ok;
+    expect(screen.getByText(/December 20, 2019 in the morning/i)).to.be.ok;
+  });
+
+  it('should render CC appointment request view', async () => {
+    const flowType = FLOW_TYPES.REQUEST;
+    const data = {
+      typeOfCareId: '323',
+      facilityType: 'communityCare',
+      distanceWillingToTravel: '25',
+      preferredLanguage: 'english',
+      visitType: 'office',
+      bestTimeToCall: {
+        morning: true,
+        afternoon: true,
+      },
+      hasCommunityCareProvider: true,
+      communityCareProvider: {
+        firstName: 'Jane',
+        lastName: 'Doe',
+        phone: '5555555555',
+        address: {
+          street: '123 Test',
+          city: 'Northampton',
+          state: 'MA',
+          postalCode: '01060',
+        },
+      },
+      calendarData: {
+        selectedDates: [{ date: '2019-12-20', optionTime: 'AM' }],
+      },
+    };
+    const facilityDetails = {
+      name: 'CHYSHR-Sidney VA Clinic',
+      id: 'var983',
+      address: {
+        postalCode: '82001-5356',
+        city: 'Cheyenne',
+        state: 'WY',
+        line: ['2360 East Pershing Boulevard'],
+      },
+    };
+
+    const screen = renderWithStoreAndRouter(
+      <ConfirmationPage
+        facilityDetails={facilityDetails}
+        flowType={flowType}
+        data={data}
+      />,
+      {
+        initialState,
+        reducers,
+      },
+    );
+
+    expect(screen.getByText(/Your appointment request has been submitted./i)).to
+      .be.ok;
+    expect(screen.getByText(/Community Care/i)).to.be.ok;
+    expect(screen.getByText(/Primary care appointment/i)).to.be.ok;
+    expect(screen.getByText(/Pending/i)).to.be.ok;
+    expect(screen.getByText(/Jane Doe/i)).to.be.ok;
+    expect(screen.getByText(/555555555/i)).to.be.ok;
+    expect(screen.getByText(/123 Test/i)).to.be.ok;
+    expect(screen.getByText(/Northampton, MA 01060/i)).to.be.ok;
+    expect(screen.getByText(/December 20, 2019 in the morning/i)).to.be.ok;
+  });
+
+  it('should render CC appointment request view without provider', () => {
+    const flowType = FLOW_TYPES.REQUEST;
+    const data = {
+      typeOfCareId: '323',
+      facilityType: 'communityCare',
+      distanceWillingToTravel: '25',
+      preferredLanguage: 'english',
+      visitType: 'office',
+      bestTimeToCall: {
+        morning: true,
+        afternoon: true,
+      },
+      hasCommunityCareProvider: false,
+      calendarData: {
+        selectedDates: [{ date: '2019-12-20', optionTime: 'AM' }],
+      },
+    };
+    const facilityDetails = {
+      name: 'CHYSHR-Sidney VA Clinic',
+      id: 'var983',
+      address: {
+        postalCode: '82001-5356',
+        city: 'Cheyenne',
+        state: 'WY',
+        line: ['2360 East Pershing Boulevard'],
+      },
+    };
+
+    const screen = renderWithStoreAndRouter(
+      <ConfirmationPage
+        facilityDetails={facilityDetails}
+        flowType={flowType}
+        data={data}
+      />,
+      {
+        initialState,
+        reducers,
+      },
+    );
+
+    expect(screen.getByText(/Your appointment request has been submitted./i)).to
+      .be.ok;
+    expect(screen.getByText(/Community Care/i)).to.be.ok;
+    expect(screen.getByText(/Primary care appointment/i)).to.be.ok;
+    expect(screen.getByText(/Pending/i)).to.be.ok;
+    expect(screen.getByText(/No preference/i)).to.be.ok;
+
+    expect(screen.queryByText(/CHYSHR-Sidney VA Clinic/i)).to.be.null;
+  });
+
+  it('should render addtional information when "Show more" button is clicked', async () => {
+    const flowType = FLOW_TYPES.REQUEST;
+    const data = {
+      typeOfCareId: '323',
+      phoneNumber: '1234567890',
+      email: 'joeblow@gmail.com',
+      reasonForAppointment: 'routine-follow-up',
+      reasonAdditionalInfo: 'Additional info',
+      bestTimeToCall: {
+        evening: true,
+        morning: false,
+        afternoon: false,
+      },
+    };
+    const facilityDetails = {
+      name: 'CHYSHR-Sidney VA Clinic',
+      id: 'var983',
+      address: {
+        postalCode: '82001-5356',
+        city: 'Cheyenne',
+        state: 'WY',
+        line: ['2360 East Pershing Boulevard'],
+      },
+    };
+
+    const screen = renderWithStoreAndRouter(
+      <ConfirmationPage
+        facilityDetails={facilityDetails}
+        flowType={flowType}
+        data={data}
+      />,
+      {
+        initialState,
+        reducers,
+      },
+    );
+
+    // Simulate user clicking the "Show more" button
+    const button = screen.getByRole('button', { name: 'Show more' });
+    userEvent.click(button);
+
+    // Button text should change to "Show less"
+    await screen.findByRole('button', { name: 'Show less' });
+
+    // The following additional information should be displayed
+    expect(screen.getByText(/Follow-up\/Routine/i)).to.be.ok;
+    expect(screen.getByText(/Additional info/i)).to.be.ok;
+    expect(screen.getByText(/joeblow@gmail.com/i)).to.be.ok;
+    expect(screen.getByText(/1234567890/i)).to.be.ok;
+    expect(screen.getByText(/Evening/i)).to.be.ok;
+  });
+
+  it('should format the best time to call correctly when 2 times are selected', async () => {
+    const flowType = FLOW_TYPES.REQUEST;
+    const data = {
+      bestTimeToCall: {
+        evening: true,
+        morning: true,
+        afternoon: false,
+      },
+    };
+    const facilityDetails = {
+      address: {},
+    };
+    const screen = renderWithStoreAndRouter(
+      <ConfirmationPage
+        facilityDetails={facilityDetails}
+        flowType={flowType}
+        data={data}
+      />,
+      {
+        initialState,
+        reducers,
+      },
+    );
+
+    // Simulate user clicking the "Show more" button
+    const button = screen.getByRole('button', { name: 'Show more' });
+    userEvent.click(button);
+
+    // Button text should change to "Show less"
+    await screen.findByRole('button', { name: 'Show less' });
+
+    expect(screen.getByText(/Morning or Evening/i)).to.be.ok;
+  });
+
+  it('should format the best time to call correctly when 3 times are selected', async () => {
+    const flowType = FLOW_TYPES.REQUEST;
+    const data = {
+      bestTimeToCall: {
+        evening: true,
+        morning: true,
+        afternoon: true,
+      },
+    };
+    const facilityDetails = {
+      address: {},
+    };
+    const screen = renderWithStoreAndRouter(
+      <ConfirmationPage
+        facilityDetails={facilityDetails}
+        flowType={flowType}
+        data={data}
+      />,
+      {
+        initialState,
+        reducers,
+      },
+    );
+
+    // Simulate user clicking the "Show more" button
+    const button = screen.getByRole('button', { name: 'Show more' });
+    userEvent.click(button);
+
+    // Button text should change to "Show less"
+    await screen.findByRole('button', { name: 'Show less' });
+
+    expect(screen.getByText(/Anytime during the day/i)).to.be.ok;
   });
 
   it('should render new appointment page when "New appointment" button is clicked', () => {
